+++ src/components/SharedComponents/CustomRangeSlider.jsx
