+++ src/pages/Store/components/Cart/index.js
@@ -1,0 +1,2 @@
+import Cart from "./Cart";
+export default Cart
